🚑️ Fix bug with circular import
