+++ aiomysql/connection.py
@@ -1,25 +1,19 @@
 # Python implementation of the MySQL client-server protocol
 # http://dev.mysql.com/doc/internals/en/client-server-protocol.html
-
+# https://github.com/PyMySQL/PyMySQL/blob/master/pymysql/cursors.py#L11-L18
 import asyncio
 import configparser
+import contextlib
 import getpass
+import json
 import os
+import re
 import socket
 import struct
 import sys
 import warnings
 from functools import partial
-from typing import (
-    Union,
-    Mapping,
-    Any,
-    Optional,
-    List,
-    Tuple,
-    Dict,
-    Type
-)
+from typing import Optional, Union
 
 from pymysql.charset import (
     charset_by_name,
@@ -67,7 +61,6 @@ from pymysql.err import (
     ProgrammingError
 )
 
-from .cursors import Cursor
 from .log import logger
 from .utils import (
     _pack_int24,
@@ -80,8 +73,718 @@ try:
 except KeyError:
     DEFAULT_USER = "unknown"
 
+#: Regular expression for :meth:`Cursor.executemany`.
+#: executemany only supports simple bulk insert.
+#: You can use it to load large dataset.
+# flake8: noqa
+RE_INSERT_VALUES = re.compile(
+    r"\s*((?:INSERT|REPLACE)\s.+\sVALUES?\s+)" +
+    r"(\(\s*(?:%s|%\(.+\)s)\s*(?:,\s*(?:%s|%\(.+\)s)\s*)*\))" +
+    r"(\s*(?:ON DUPLICATE.*)?);?\s*\Z",
+    re.IGNORECASE | re.DOTALL)
 
-async def connect(
+
+class Cursor:
+    """Cursor is used to interact with the database."""
+
+    #: Max statement size which :meth:`executemany` generates.
+    #:
+    #: Max size of allowed statement is max_allowed_packet -
+    # packet_header_size.
+    #: Default value of max_allowed_packet is 1048576.
+    max_stmt_length = 1024000
+
+    def __init__(self, connection, echo=False):
+        """
+        Do not create an instance of a Cursor yourself. Call
+        connections.Connection.cursor().
+        """
+        self._connection = connection
+        self._loop = self._connection.loop
+        self._description = None
+        self._rownumber = 0
+        self._rowcount = -1
+        self._arraysize = 1
+        self._executed = None
+        self._result = None
+        self._rows = None
+        self._lastrowid = None
+        self._echo = echo
+
+    @property
+    def connection(self):
+        """
+        This read-only attribute return a reference to the Connection
+        object on which the cursor was created.
+        """
+        return self._connection
+
+    @property
+    def description(self):
+        """
+        This read-only attribute is a sequence of 7-item sequences.
+
+        Each of these sequences is a collections.namedtuple containing
+        information describing one result column:
+
+        0.  name: the name of the column returned.
+        1.  type_code: the type of the column.
+        2.  display_size: the actual length of the column in bytes.
+        3.  internal_size: the size in bytes of the column associated to
+            this column on the server.
+        4.  precision: total number of significant digits in columns of
+            type NUMERIC. None for other types.
+        5.  scale: count of decimal digits in the fractional part in
+            columns of type NUMERIC. None for other types.
+        6.  null_ok: always None as not easy to retrieve from the libpq.
+
+        This attribute will be None for operations that do not
+        return rows or if the cursor has not had an operation invoked
+        via the execute() method yet.
+        """
+        return self._description
+
+    @property
+    def rowcount(self):
+        """
+        Returns the number of rows that has been produced of affected.
+
+        This read-only attribute specifies the number of rows that the
+        last :meth:`execute` produced (for Data Query Language
+        statements like SELECT) or affected (for Data Manipulation
+        Language statements like UPDATE or INSERT).
+
+        The attribute is -1 in case no .execute() has been performed
+        on the cursor or the row count of the last operation if it
+        can't be determined by the interface.
+        """
+        return self._rowcount
+
+    @property
+    def rownumber(self):
+        """
+        Row index.
+
+        This read-only attribute provides the current 0-based index of the
+        cursor in the result set or ``None`` if the index cannot be
+        determined.
+        """
+
+        return self._rownumber
+
+    @property
+    def arraysize(self):
+        """
+        How many rows will be returned by fetchmany() call.
+
+        This read/write attribute specifies the number of rows to
+        fetch at a time with fetchmany(). It defaults to
+        1 meaning to fetch a single row at a time.
+
+        """
+        return self._arraysize
+
+    @arraysize.setter
+    def arraysize(self, val):
+        """
+        How many rows will be returned by fetchmany() call.
+
+        This read/write attribute specifies the number of rows to
+        fetch at a time with fetchmany(). It defaults to
+        1 meaning to fetch a single row at a time.
+
+        """
+        self._arraysize = val
+
+    @property
+    def lastrowid(self):
+        """
+        This read-only property returns the value generated for an
+        AUTO_INCREMENT column by the previous INSERT or UPDATE statement
+        or None when there is no such value available. For example,
+        if you perform an INSERT into a table that contains an AUTO_INCREMENT
+        column, lastrowid returns the AUTO_INCREMENT value for the new row.
+        """
+        return self._lastrowid
+
+    @property
+    def echo(self):
+        """Return echo mode status."""
+        return self._echo
+
+    @property
+    def closed(self):
+        """
+        The readonly property that returns ``True`` if connections was
+        detached from current cursor
+        """
+        return True if not self._connection else False
+
+    async def close(self):
+        """Closing a cursor just exhausts all remaining data."""
+        conn = self._connection
+        if conn is None:
+            return
+        try:
+            while await self.nextset():
+                pass
+        finally:
+            self._connection = None
+
+    def _get_db(self):
+        if not self._connection:
+            raise ProgrammingError("Cursor closed")
+        return self._connection
+
+    def _check_executed(self):
+        if not self._executed:
+            raise ProgrammingError("execute() first")
+
+    def _conv_row(self, row):
+        return row
+
+    def setinputsizes(self, *args):
+        """Does nothing, required by DB API."""
+
+    def setoutputsizes(self, *args):
+        """Does nothing, required by DB API."""
+
+    async def nextset(self):
+        """Get the next query set"""
+        conn = self._get_db()
+        current_result = self._result
+        if current_result is None or current_result is not conn._result:
+            return
+        if not current_result.has_next:
+            return
+        self._result = None
+        self._clear_result()
+        await conn.next_result()
+        await self._do_get_result()
+        return True
+
+    def _escape_args(self, args, conn):
+        if isinstance(args, (tuple, list)):
+            return tuple(conn.escape(arg) for arg in args)
+        elif isinstance(args, dict):
+            return dict((key, conn.escape(val)) for (key, val) in args.items())
+        else:
+            # If it's not a dictionary let's try escaping it anyway.
+            # Worst case it will throw a Value error
+            return conn.escape(args)
+
+    def mogrify(self, query, args=None):
+        """
+        Returns the exact string that is sent to the database by calling
+        to execute() method. This method follows the extension to the DB
+        API 2.0 followed by Psycopg.
+
+        :param query: ``str`` sql statement
+        :param args: ``tuple`` or ``list`` of arguments for sql query
+        """
+        conn = self._get_db()
+        if args is not None:
+            query = query % self._escape_args(args, conn)
+        return query
+
+    async def execute(self, query, args=None):
+        """
+        Executes the given operation
+
+        Executes the given operation substituting any markers with
+        the given parameters.
+
+        For example, getting all rows where id is 5:
+          cursor.execute("SELECT * FROM t1 WHERE id = %s", (5,))
+
+        :param query: ``str`` sql statement
+        :param args: ``tuple`` or ``list`` of arguments for sql query
+        :returns: ``int``, number of rows that has been produced of affected
+        """
+        conn = self._get_db()
+
+        while await self.nextset():
+            pass
+
+        if args is not None:
+            query = query % self._escape_args(args, conn)
+
+        await self._query(query)
+        self._executed = query
+        if self._echo:
+            logger.info(query)
+            logger.info("%r", args)
+        return self._rowcount
+
+    async def executemany(self, query, args):
+        """
+        Execute the given operation multiple times
+
+        The executemany() method will execute the operation iterating
+        over the list of parameters in seq_params.
+
+        Example: Inserting 3 new employees and their phone number
+
+            data = [
+                ('Jane','555-001'),
+                ('Joe', '555-001'),
+                ('John', '555-003')
+                ]
+            stmt = "INSERT INTO employees (name, phone) VALUES ('%s','%s')"
+            await cursor.executemany(stmt, data)
+
+        INSERT or REPLACE statements are optimized by batching the data,
+        that is using the MySQL multiple rows syntax.
+
+        :param query: `str`, sql statement
+        :param args: ``tuple`` or ``list`` of arguments for sql query
+        """
+        if not args:
+            return
+
+        if self._echo:
+            logger.info("CALL %s", query)
+            logger.info("%r", args)
+
+        m = RE_INSERT_VALUES.match(query)
+        if m:
+            q_prefix = m.group(1) % ()
+            q_values = m.group(2).rstrip()
+            q_postfix = m.group(3) or ''
+            assert q_values[0] == '(' and q_values[-1] == ')'
+            return (await self._do_execute_many(
+                q_prefix, q_values, q_postfix, args, self.max_stmt_length,
+                self._get_db().encoding))
+        else:
+            rows = 0
+            for arg in args:
+                await self.execute(query, arg)
+                rows += self._rowcount
+            self._rowcount = rows
+        return self._rowcount
+
+    async def _do_execute_many(self, prefix, values, postfix, args,
+                               max_stmt_length, encoding):
+        conn = self._get_db()
+        escape = self._escape_args
+        if isinstance(prefix, str):
+            prefix = prefix.encode(encoding)
+        if isinstance(postfix, str):
+            postfix = postfix.encode(encoding)
+        sql = bytearray(prefix)
+        args = iter(args)
+        v = values % escape(next(args), conn)
+        if isinstance(v, str):
+            v = v.encode(encoding, 'surrogateescape')
+        sql += v
+        rows = 0
+        for arg in args:
+            v = values % escape(arg, conn)
+            if isinstance(v, str):
+                v = v.encode(encoding, 'surrogateescape')
+            if len(sql) + len(v) + len(postfix) + 1 > max_stmt_length:
+                r = await self.execute(sql + postfix)
+                rows += r
+                sql = bytearray(prefix)
+            else:
+                sql += b','
+            sql += v
+        r = await self.execute(sql + postfix)
+        rows += r
+        self._rowcount = rows
+        return rows
+
+    async def callproc(self, procname, args=()):
+        """
+        Execute stored procedure procname with args
+
+        Compatibility warning: PEP-249 specifies that any modified
+        parameters must be returned. This is currently impossible
+        as they are only available by storing them in a server
+        variable and then retrieved by a query. Since stored
+        procedures return zero or more result sets, there is no
+        reliable way to get at OUT or INOUT parameters via callproc.
+        The server variables are named @_procname_n, where procname
+        is the parameter above and n is the position of the parameter
+        (from zero). Once all result sets generated by the procedure
+        have been fetched, you can issue a SELECT @_procname_0, ...
+        query using .execute() to get any OUT or INOUT values.
+
+        Compatibility warning: The act of calling a stored procedure
+        itself creates an empty result set. This appears after any
+        result sets generated by the procedure. This is non-standard
+        behavior with respect to the DB-API. Be sure to use nextset()
+        to advance through all result sets; otherwise you may get
+        disconnected.
+
+        :param procname: ``str``, name of procedure to execute on server
+        :param args: `sequence of parameters to use with procedure
+        :returns: the original args.
+        """
+        conn = self._get_db()
+        if self._echo:
+            logger.info("CALL %s", procname)
+            logger.info("%r", args)
+
+        for index, arg in enumerate(args):
+            q = "SET @_%s_%d=%s" % (procname, index, conn.escape(arg))
+            await self._query(q)
+            await self.nextset()
+
+        _args = ','.join('@_%s_%d' % (procname, i) for i in range(len(args)))
+        q = "CALL %s(%s)" % (procname, _args)
+        await self._query(q)
+        self._executed = q
+        return args
+
+    def fetchone(self):
+        """Fetch the next row """
+        self._check_executed()
+        fut = self._loop.create_future()
+
+        if self._rows is None or self._rownumber >= len(self._rows):
+            fut.set_result(None)
+            return fut
+        result = self._rows[self._rownumber]
+        self._rownumber += 1
+
+        fut = self._loop.create_future()
+        fut.set_result(result)
+        return fut
+
+    def fetchmany(self, size=None):
+        """
+        Returns the next set of rows of a query result, returning a
+        list of tuples. When no more rows are available, it returns an
+        empty list.
+
+        The number of rows returned can be specified using the size argument,
+        which defaults to one
+
+        :param size: ``int`` number of rows to return
+        :returns: ``list`` of fetched rows
+        """
+        self._check_executed()
+        fut = self._loop.create_future()
+        if self._rows is None:
+            fut.set_result([])
+            return fut
+        end = self._rownumber + (size or self._arraysize)
+        result = self._rows[self._rownumber:end]
+        self._rownumber = min(end, len(self._rows))
+
+        fut.set_result(result)
+        return fut
+
+    def fetchall(self):
+        """
+        Returns all rows of a query result set
+
+        :returns: ``list`` of fetched rows
+        """
+        self._check_executed()
+        fut = self._loop.create_future()
+        if self._rows is None:
+            fut.set_result([])
+            return fut
+
+        if self._rownumber:
+            result = self._rows[self._rownumber:]
+        else:
+            result = self._rows
+        self._rownumber = len(self._rows)
+
+        fut.set_result(result)
+        return fut
+
+    def scroll(self, value, mode='relative'):
+        """Scroll the cursor in the result set to a new position according
+         to mode.
+
+        If mode is relative (default), value is taken as offset to the
+        current position in the result set, if set to absolute, value
+        states an absolute target position. An IndexError should be raised in
+        case a scroll operation would leave the result set. In this case,
+        the cursor position is left undefined (ideal would be to
+        not move the cursor at all).
+
+        :param int value: move cursor to next position according to mode.
+        :param str mode: scroll mode, possible modes: `relative` and `absolute`
+        """
+        self._check_executed()
+        if mode == 'relative':
+            r = self._rownumber + value
+        elif mode == 'absolute':
+            r = value
+        else:
+            raise ProgrammingError("unknown scroll mode %s" % mode)
+
+        if not (0 <= r < len(self._rows)):
+            raise IndexError("out of range")
+        self._rownumber = r
+
+        fut = self._loop.create_future()
+        fut.set_result(None)
+        return fut
+
+    async def _query(self, q):
+        conn = self._get_db()
+        self._last_executed = q
+        self._clear_result()
+        await conn.query(q)
+        await self._do_get_result()
+
+    def _clear_result(self):
+        self._rownumber = 0
+        self._result = None
+
+        self._rowcount = 0
+        self._description = None
+        self._lastrowid = None
+        self._rows = None
+
+    async def _do_get_result(self):
+        conn = self._get_db()
+        self._rownumber = 0
+        self._result = result = conn._result
+        self._rowcount = result.affected_rows
+        self._description = result.description
+        self._lastrowid = result.insert_id
+        self._rows = result.rows
+
+        if result.warning_count > 0:
+            await self._show_warnings(conn)
+
+    async def _show_warnings(self, conn):
+        if self._result and self._result.has_next:
+            return
+        ws = await conn.show_warnings()
+        if ws is None:
+            return
+        for w in ws:
+            msg = w[-1]
+            warnings.warn(str(msg), Warning, 4)
+
+    Warning = Warning
+    Error = Error
+    InterfaceError = InterfaceError
+    DatabaseError = DatabaseError
+    DataError = DataError
+    OperationalError = OperationalError
+    IntegrityError = IntegrityError
+    InternalError = InternalError
+    ProgrammingError = ProgrammingError
+    NotSupportedError = NotSupportedError
+
+    def __aiter__(self):
+        return self
+
+    async def __anext__(self):
+        ret = await self.fetchone()
+        if ret is not None:
+            return ret
+        else:
+            raise StopAsyncIteration  # noqa
+
+    async def __aenter__(self):
+        return self
+
+    async def __aexit__(self, exc_type, exc_val, exc_tb):
+        await self.close()
+        return
+
+
+class _DeserializationCursorMixin:
+    async def _do_get_result(self):
+        await super()._do_get_result()
+        if self._rows:
+            self._rows = [self._deserialization_row(r) for r in self._rows]
+
+    def _deserialization_row(self, row):
+        if row is None:
+            return None
+        if isinstance(row, dict):
+            dict_flag = True
+        else:
+            row = list(row)
+            dict_flag = False
+        for index, (name, field_type, *n) in enumerate(self._description):
+            if field_type == FIELD_TYPE.JSON:
+                point = name if dict_flag else index
+                with contextlib.suppress(ValueError, TypeError):
+                    row[point] = json.loads(row[point])
+        if dict_flag:
+            return row
+        else:
+            return tuple(row)
+
+    def _conv_row(self, row):
+        if row is None:
+            return None
+        row = super()._conv_row(row)
+        return self._deserialization_row(row)
+
+
+class DeserializationCursor(_DeserializationCursorMixin, Cursor):
+    """A cursor automatic deserialization of json type fields"""
+
+
+class _DictCursorMixin:
+    # You can override this to use OrderedDict or other dict-like types.
+    dict_type = dict
+
+    async def _do_get_result(self):
+        await super()._do_get_result()
+        fields = []
+        if self._description:
+            for f in self._result.fields:
+                name = f.name
+                if name in fields:
+                    name = f.table_name + '.' + name
+                fields.append(name)
+            self._fields = fields
+
+        if fields and self._rows:
+            self._rows = [self._conv_row(r) for r in self._rows]
+
+    def _conv_row(self, row):
+        if row is None:
+            return None
+        row = super()._conv_row(row)
+        return self.dict_type(zip(self._fields, row))
+
+
+class DictCursor(_DictCursorMixin, Cursor):
+    """A cursor which returns results as a dictionary"""
+
+
+class SSCursor(Cursor):
+    """
+    Unbuffered Cursor, mainly useful for queries that return a lot of
+    data, or for connections to remote servers over a slow network.
+
+    Instead of copying every row of data into a buffer, this will fetch
+    rows as needed. The upside of this, is the client uses much less memory,
+    and rows are returned much faster when traveling over a slow network,
+    or if the result set is very big.
+
+    There are limitations, though. The MySQL protocol doesn't support
+    returning the total number of rows, so the only way to tell how many rows
+    there are is to iterate over every row returned. Also, it currently isn't
+    possible to scroll backwards, as only the current row is held in memory.
+    """
+
+    async def close(self):
+        conn = self._connection
+        if conn is None:
+            return
+
+        if self._result is not None and self._result is conn._result:
+            await self._result._finish_unbuffered_query()
+
+        try:
+            while await self.nextset():
+                pass
+        finally:
+            self._connection = None
+
+    async def _query(self, q):
+        conn = self._get_db()
+        self._last_executed = q
+        await conn.query(q, unbuffered=True)
+        await self._do_get_result()
+        return self._rowcount
+
+    async def _read_next(self):
+        """Read next row """
+        row = await self._result._read_rowdata_packet_unbuffered()
+        row = self._conv_row(row)
+        return row
+
+    async def fetchone(self):
+        """ Fetch next row """
+        self._check_executed()
+        row = await self._read_next()
+        if row is None:
+            return
+        self._rownumber += 1
+        return row
+
+    async def fetchall(self):
+        """Fetch all, as per MySQLdb. Pretty useless for large queries, as
+        it is buffered.
+        """
+        rows = []
+        while True:
+            row = await self.fetchone()
+            if row is None:
+                break
+            rows.append(row)
+        return rows
+
+    async def fetchmany(self, size=None):
+        """Returns the next set of rows of a query result, returning a
+        list of tuples. When no more rows are available, it returns an
+        empty list.
+
+        The number of rows returned can be specified using the size argument,
+        which defaults to one
+
+        :param size: ``int`` number of rows to return
+        :returns: ``list`` of fetched rows
+        """
+        self._check_executed()
+        if size is None:
+            size = self._arraysize
+
+        rows = []
+        for i in range(size):
+            row = await self._read_next()
+            if row is None:
+                break
+            rows.append(row)
+            self._rownumber += 1
+        return rows
+
+    async def scroll(self, value, mode='relative'):
+        """Scroll the cursor in the result set to a new position
+        according to mode . Same as :meth:`Cursor.scroll`, but move cursor
+        on server side one by one row. If you want to move 20 rows forward
+        scroll will make 20 queries to move cursor. Currently, only forward
+        scrolling is supported.
+
+        :param int value: move cursor to next position according to mode.
+        :param str mode: scroll mode, possible modes: `relative` and `absolute`
+        """
+
+        self._check_executed()
+
+        if mode == 'relative':
+            if value < 0:
+                raise NotSupportedError("Backwards scrolling not supported "
+                                        "by this cursor")
+
+            for _ in range(value):
+                await self._read_next()
+            self._rownumber += value
+        elif mode == 'absolute':
+            if value < self._rownumber:
+                raise NotSupportedError(
+                    "Backwards scrolling not supported by this cursor")
+
+            end = value - self._rownumber
+            for _ in range(end):
+                await self._read_next()
+            self._rownumber = value
+        else:
+            raise ProgrammingError(f"unknown scroll {mode}")
+
+
+class SSDictCursor(_DictCursorMixin, SSCursor):
+    """An unbuffered cursor, which returns results as a dictionary """
+
+
+def connect(
         host: str = "localhost",
         user: Optional[str] = None,
         password: str = "",
@@ -91,10 +794,10 @@ async def connect(
         charset: str = '',
         sql_mode: Optional[str] = None,
         read_default_file: Optional[str] = None,
-        conv: Optional[Dict[str, Any]] = None,
-        use_unicode: Optional[bool] = None,
+        conv: Optional[dict] = decoders,
+        use_unicode: Optional[Union[bool, str]] = None,
         client_flag: int = 0,
-        cursorclass: Type[Cursor] = Cursor,
+        cursorclass: "Cursor" = Cursor,
         init_command: Optional[str] = None,
         connect_timeout: Optional[int] = None,
         read_default_group: Optional[str] = None,
@@ -102,50 +805,61 @@ async def connect(
         echo: bool = False,
         local_infile: bool = False,
         loop: Optional[asyncio.AbstractEventLoop] = None,
-        ssl: Optional[Union[bool, Dict[str, Any]]] = None,
-        auth_plugin: Optional[str] = '',
-        program_name: Optional[str] = '',
-        server_public_key: Optional[Union[str, bytes]] = None,
-) -> 'Connection':
+        ssl: Optional[dict] = None,
+        auth_plugin: str = '',
+        program_name: str = '',
+        server_public_key: Optional[bytes] = None
+) -> _ContextManager["Connection"]:
     """See connections.Connection.__init__() for information about
     defaults."""
-    coro = _connect(host=host, user=user, password=password, db=db,
-                    port=port, unix_socket=unix_socket, charset=charset,
-                    sql_mode=sql_mode, read_default_file=read_default_file,
-                    conv=conv, use_unicode=use_unicode,
-                    client_flag=client_flag, cursorclass=cursorclass,
-                    init_command=init_command,
-                    connect_timeout=connect_timeout,
-                    read_default_group=read_default_group,
-                    autocommit=autocommit, echo=echo,
-                    local_infile=local_infile, loop=loop, ssl=ssl,
-                    auth_plugin=auth_plugin, program_name=program_name)
-    return _ContextManager[Connection](coro, disconnect)
+    coro = _connect(
+        host=host,
+        user=user,
+        password=password,
+        db=db,
+        port=port,
+        unix_socket=unix_socket,
+        charset=charset,
+        sql_mode=sql_mode,
+        read_default_file=read_default_file,
+        conv=conv,
+        use_unicode=use_unicode,
+        client_flag=client_flag,
+        cursorclass=cursorclass,
+        init_command=init_command,
+        connect_timeout=connect_timeout,
+        read_default_group=read_default_group,
+        autocommit=autocommit,
+        echo=echo,
+        local_infile=local_infile,
+        loop=loop,
+        ssl=ssl,
+        auth_plugin=auth_plugin,
+        program_name=program_name
+    )
+    return _ContextManager[Connection](coro, _disconnect)
 
 
-async def disconnect(c: "Connection") -> None:
-    c.close()
+# noinspection PyUnresolvedReferences
+async def _disconnect(c: "Connection"):
+    await c.close()
 
 
-# todo: Update Any to stricter kwarg
-# https://github.com/python/mypy/issues/4441
+async def _close_cursor(c: Cursor):
+    await c.close()
+
+
+# noinspection PyProtectedMember
 async def _connect(*args, **kwargs):
     conn = Connection(*args, **kwargs)
     await conn._connect()
     return conn
 
 
-# todo: Update Any to stricter kwarg
-# https://github.com/python/mypy/issues/4441
-async def _open_connection(
-        host: Optional[str] = None,
-        port: Optional[int] = None,
-        **kwds: Any
-) -> Tuple['_StreamReader', asyncio.StreamWriter]:
-    """This is based on asyncio.open_connection, allowing us to use a custom
+async def _open_connection(host=None, port=None, **kwds):
+    """
+    This is based on asyncio.open_connection, allowing us to use a custom
     StreamReader.
-
-    `limit` arg has been removed as we don't currently use it.
     """
     loop = asyncio.events.get_running_loop()
     reader = _StreamReader(loop=loop)
@@ -156,18 +870,12 @@ async def _open_connection(
     return reader, writer
 
 
-# todo: Update Any to stricter kwarg
-# https://github.com/python/mypy/issues/4441
-async def _open_unix_connection(
-        path: Optional[str] = None,
-        **kwds: Any
-) -> Tuple['_StreamReader', asyncio.StreamWriter]:
-    """This is based on asyncio.open_unix_connection, allowing us to use a custom
-    StreamReader.
-
-    `limit` arg has been removed as we don't currently use it.
+async def _open_unix_connection(path=None, **kwds):
     """
-    loop = asyncio.get_running_loop()
+    This is based on asyncio.open_unix_connection, allowing us to use a custom
+    StreamReader.
+    """
+    loop = asyncio.events.get_running_loop()
 
     reader = _StreamReader(loop=loop)
     protocol = asyncio.StreamReaderProtocol(reader, loop=loop)
@@ -178,15 +886,14 @@ async def _open_unix_connection(
 
 
 class _StreamReader(asyncio.StreamReader):
-    """This StreamReader exposes whether EOF was received, allowing us to
+    """
+    This StreamReader exposes whether EOF was received, allowing us to
     discard the associated connection instead of returning it from the pool
     when checking free connections in Pool._fill_free_pool().
-
-    `limit` arg has been removed as we don't currently use it.
     """
 
-    def __init__(self, loop: Optional[asyncio.AbstractEventLoop] = None) -> None:
-        self._eof_received: bool = False
+    def __init__(self, loop=None):
+        self._eof_received = False
         super().__init__(loop=loop)
 
     def feed_eof(self) -> None:
@@ -194,12 +901,13 @@ class _StreamReader(asyncio.StreamReader):
         super().feed_eof()
 
     @property
-    def eof_received(self) -> bool:
+    def eof_received(self):
         return self._eof_received
 
 
 class Connection:
-    """Representation of a socket with a mysql server.
+    """
+    Representation of a socket with a mysql server.
 
     The proper way to get an instance of this class is to call
     connect().
@@ -207,31 +915,31 @@ class Connection:
 
     def __init__(
             self,
-            host: str = "localhost",
-            user: Optional[str] = None,
-            password: str = "",
-            db: Optional[str] = None,
-            port: int = 3306,
-            unix_socket: Optional[str] = None,
-            charset: str = '',
-            sql_mode: Optional[str] = None,
-            read_default_file: Optional[str] = None,
-            conv: dict = decoders,
-            use_unicode: Optional[bool] = None,
-            client_flag: int = 0,
-            cursorclass: type = Cursor,
-            init_command: Optional[str] = None,
-            connect_timeout: Optional[int] = None,
-            read_default_group: Optional[str] = None,
-            autocommit: Optional[bool] = False,
-            echo: Optional[bool] = False,
-            local_infile: Optional[bool] = False,
-            loop: Optional[asyncio.AbstractEventLoop] = None,
-            ssl: Optional[Union[str, Mapping[str, Any]]] = None,
-            auth_plugin: Optional[str] = '',
-            program_name: Optional[str] = '',
-            server_public_key: Optional[str] = None
-    ) -> None:
+            host="localhost",
+            user=None,
+            password="",
+            db=None,
+            port=3306,
+            unix_socket=None,
+            charset='',
+            sql_mode=None,
+            read_default_file=None,
+            conv=decoders,
+            use_unicode=None,
+            client_flag=0,
+            cursorclass=Cursor,
+            init_command=None,
+            connect_timeout=None,
+            read_default_group=None,
+            autocommit=False,
+            echo=False,
+            local_infile=False,
+            loop=None,
+            ssl=None,
+            auth_plugin='',
+            program_name='',
+            server_public_key=None
+    ):
         """
         Establish a connection to the MySQL database. Accepts several
         arguments:
@@ -367,69 +1075,70 @@ class Connection:
         self._close_reason = None
 
     @property
-    def host(self) -> str:
+    def host(self):
         """MySQL server IP address or name"""
         return self._host
 
     @property
-    def port(self) -> int:
+    def port(self):
         """MySQL server TCP/IP port"""
         return self._port
 
     @property
-    def unix_socket(self) -> Optional[str]:
+    def unix_socket(self):
         """MySQL Unix socket file location"""
         return self._unix_socket
 
     @property
-    def db(self) -> Optional[str]:
+    def db(self):
         """Current database name."""
         return self._db
 
     @property
-    def user(self) -> Optional[str]:
+    def user(self):
         """User used while connecting to MySQL"""
         return self._user
 
     @property
-    def echo(self) -> Optional[bool]:
+    def echo(self):
         """Return echo mode status."""
         return self._echo
 
     @property
-    def last_usage(self) -> Optional[float]:
+    def last_usage(self):
         """Return time() when connection was used."""
         return self._last_usage
 
     @property
-    def loop(self) -> Optional[asyncio.AbstractEventLoop]:
+    def loop(self):
         return self._loop
 
     @property
-    def closed(self) -> Optional[bool]:
-        """The readonly property that returns ``True`` if connections is
+    def closed(self):
+        """
+        The readonly property that returns ``True`` if connections is
         closed.
         """
         return self._writer is None
 
     @property
-    def encoding(self) -> str:
+    def encoding(self):
         """Encoding employed for this connection."""
         return self._encoding
 
     @property
-    def charset(self) -> str:
+    def charset(self):
         """Returns the character set for current connection."""
         return self._charset
 
-    def close(self) -> None:
+    def close(self):
         """Close socket connection"""
         if self._writer:
             self._writer.transport.close()
         self._writer = None
         self._reader = None
 
-    async def ensure_closed(self) -> None:
+    async def ensure_closed(self):
         """Send quit command and then close socket connection"""
         if self._writer is None:
             # connection has been closed
@@ -439,8 +1148,9 @@ class Connection:
         await self._writer.drain()
         self.close()
 
-    async def autocommit(self, value) -> None:
-        """Enable/disable autocommit mode for current MySQL session.
+    async def autocommit(self, value):
+        """
+        Enable/disable autocommit mode for current MySQL session.
 
         :param value: ``bool``, toggle autocommit
         """
@@ -450,14 +1160,12 @@ class Connection:
             await self._send_autocommit_mode()
 
     def get_autocommit(self) -> bool:
-        """Returns autocommit status for current MySQL session.
-
-        :returns bool: current autocommit status."""
+        """Returns autocommit status for current MySQL session."""
 
         status = self.server_status & SERVER_STATUS.SERVER_STATUS_AUTOCOMMIT
         return bool(status)
 
-    async def _read_ok_packet(self) -> bool:
+    async def _read_ok_packet(self):
         pkt = await self._read_packet()
         if not pkt.is_ok_packet():
             raise OperationalError(2014, "Command Out of Sync")
@@ -465,65 +1173,67 @@ class Connection:
         self.server_status = ok.server_status
         return True
 
-    async def _send_autocommit_mode(self) -> None:
+    async def _send_autocommit_mode(self):
         """Set whether to commit after every execute() """
         await self._execute_command(
             COMMAND.COM_QUERY,
             "SET AUTOCOMMIT = %s" % self.escape(self.autocommit_mode))
         await self._read_ok_packet()
 
-    async def begin(self) -> None:
+    async def begin(self):
         """Begin transaction."""
         await self._execute_command(COMMAND.COM_QUERY, "BEGIN")
         await self._read_ok_packet()
 
-    async def commit(self) -> None:
+    async def commit(self):
         """Commit changes to stable storage."""
         await self._execute_command(COMMAND.COM_QUERY, "COMMIT")
         await self._read_ok_packet()
 
-    async def rollback(self) -> None:
+    async def rollback(self):
         """Roll back the current transaction."""
         await self._execute_command(COMMAND.COM_QUERY, "ROLLBACK")
         await self._read_ok_packet()
 
-    async def select_db(self, db) -> None:
+    async def select_db(self, db):
         """Set current db"""
         await self._execute_command(COMMAND.COM_INIT_DB, db)
         await self._read_ok_packet()
 
-    async def show_warnings(self) -> List[Tuple[int, str, Optional[int]]]:
+    async def show_warnings(self):
         """SHOW WARNINGS"""
         await self._execute_command(COMMAND.COM_QUERY, "SHOW WARNINGS")
         result = MySQLResult(self)
         await result.read()
         return result.rows
 
-    def escape(self, obj: Union[str, bytes, Any]) -> str:
-        """ Escape whatever value you pass to it"""
+    def escape(self, obj):
+        """Escape whatever value you pass to it"""
         if isinstance(obj, str):
             return "'" + self.escape_string(obj) + "'"
         if isinstance(obj, bytes):
             return escape_bytes_prefixed(obj)
         return escape_item(obj, self._charset)
 
-    def literal(self, obj: Union[str, bytes, Any]) -> str:
+    def literal(self, obj):
         """Alias for escape()"""
         return self.escape(obj)
 
-    def escape_string(self, s: str) -> str:
-        if self.server_status & SERVER_STATUS.SERVER_STATUS_NO_BACKSLASH_ESCAPES:
+    def escape_string(self, s):
+        if (self.server_status &
+                SERVER_STATUS.SERVER_STATUS_NO_BACKSLASH_ESCAPES):
             return s.replace("'", "''")
         return escape_string(s)
 
     def cursor(self, *cursors):
-        """Instantiates and returns a cursor
+        """
+        Instantiates and returns a cursor
 
         By default, :class:`Cursor` is returned. It is possible to also give a
         custom cursor through the cursor_class parameter, but it needs to
         be a subclass  of :class:`Cursor`
 
-        :param cursor: custom cursor class.
+        :param cursors: custom cursor class.
         :returns: instance of cursor, by default :class:`Cursor`
         :raises TypeError: cursor_class is not a subclass of Cursor.
         """
@@ -538,7 +1248,8 @@ class Connection:
         if cursors and len(cursors) == 1:
             cur = cursors[0](self, self._echo)
         elif cursors:
-            cursor_name = ''.join(map(lambda x: x.__name__, cursors)).replace('Cursor', '') + 'Cursor'
+            cursor_name = ''.join(map(lambda x: x.__name__, cursors)) \
+                              .replace('Cursor', '') + 'Cursor'
             cursor_class = type(cursor_name, cursors, {})
             cur = cursor_class(self, self._echo)
         else:
@@ -579,12 +1290,12 @@ class Connection:
         try:
             await self._execute_command(COMMAND.COM_PING, "")
             await self._read_ok_packet()
-        except Exception:
+        except Exception as ex:
             if reconnect:
                 await self._connect()
                 await self.ping(False)
             else:
-                raise
+                raise Error(f"Ping failed and cannot reconnect: {ex}")
 
     async def set_charset(self, charset):
         """Sets the character set for the current connection"""
@@ -652,10 +1363,11 @@ class Connection:
                     "Can't connect to MySQL server on %r" % self._host,
                 ) from e
 
-            # If e is neither IOError nor OSError, it's a bug.
-            # Raising AssertionError would hide the original error, so we just
-            # reraise it.
-            raise
+            # Raise an error for any other exceptions that occurred during the connection process
+            raise OperationalError(
+                CR.CR_CONN_HOST_ERROR,
+                "Encountered an error while connecting to MySQL server on %r: %r" % (self._host, e),
+            ) from e
 
     def _set_keep_alive(self):
         transport = self._writer.transport
@@ -677,8 +1389,9 @@ class Connection:
         transport.resume_reading()
 
     def write_packet(self, payload):
-        """Writes an entire "mysql packet" in its entirety to the network
-        addings its length and sequence number.
+        """
+        Writes an entire "mysql packet" in its entirety to the network
+        adding its length and sequence number.
         """
         # Internal note: when you build packet manually and calls
         # _write_bytes() directly, you should set self._next_seq_id properly.
@@ -687,7 +1400,8 @@ class Connection:
         self._next_seq_id = (self._next_seq_id + 1) % 256
 
     async def _read_packet(self, packet_type=MysqlPacket):
-        """Read an entire "mysql packet" in its entirety from the network
+        """
+        Read an entire "mysql packet" in its entirety from the network
         and return a MysqlPacket type that represents the results.
         """
         buff = b''
@@ -755,24 +1469,23 @@ class Connection:
     def _write_bytes(self, data):
         return self._writer.write(data)
 
-    async def _read_query_result(self, unbuffered: bool = False) -> None:
-        try:
-            if unbuffered:
+    async def _read_query_result(self, unbuffered=False):
+        result = None
+        if unbuffered:
+            try:
                 result = MySQLResult(self)
                 await result.init_unbuffered_query()
-            else:
-                result = MySQLResult(self)
-                await result.read()
-
-            self._result = result
-            self._affected_rows = result.affected_rows
-            self.server_status = getattr(result, 'server_status', None)
-
-        except Exception:
-            self._result = None
-            self._affected_rows = -1
-            self.server_status = None
-            raise
+            except BaseException:
+                result.unbuffered_active = False
+                result.connection = None
+                raise
+        else:
+            result = MySQLResult(self)
+            await result.read()
+        self._result = result
+        self._affected_rows = result.affected_rows
+        if result.server_status is not None:
+            self.server_status = result.server_status
 
     def insert_id(self):
         if self._result:
@@ -798,6 +1511,7 @@ class Connection:
         if self._result is not None:
             if self._result.unbuffered_active:
                 warnings.warn("Previous unbuffered result was left incomplete")
+                # noinspection PyProtectedMember
                 await self._result._finish_unbuffered_query()
             while self._result.has_next:
                 await self.next_result()
@@ -895,6 +1609,7 @@ class Connection:
         elif auth_plugin in ('', 'mysql_clear_password'):
             authresp = self._password.encode('latin1') + b'\0'
 
+        # noinspection PyUnresolvedReferences
         if self.server_capabilities & CLIENT.PLUGIN_AUTH_LENENC_CLIENT_DATA:
             data += _lenenc_int(len(authresp)) + authresp
         elif self.server_capabilities & CLIENT.SECURE_CONNECTION:
@@ -910,7 +1625,7 @@ class Connection:
             else:
                 db = self._db
             data += db + b'\0'
-
+        # noinspection PyUnresolvedReferences
         if self.server_capabilities & CLIENT.PLUGIN_AUTH:
             name = auth_plugin
             if isinstance(name, str):
@@ -920,6 +1635,7 @@ class Connection:
         self._auth_plugin_used = auth_plugin
 
         # Sends the server a few pieces of client info
+        # noinspection PyUnresolvedReferences
         if self.server_capabilities & CLIENT.CONNECT_ATTRS:
             connect_attrs = b''
             for k, v in self._connect_attrs.items():
@@ -938,7 +1654,9 @@ class Connection:
             # connection-phase-packets.html#packet-Protocol::AuthSwitchRequest
             auth_packet.read_uint8()  # 0xfe packet identifier
             plugin_name = auth_packet.read_string()
-            if self.server_capabilities & CLIENT.PLUGIN_AUTH and plugin_name is not None:
+            # noinspection PyUnresolvedReferences
+            if (self.server_capabilities & CLIENT.PLUGIN_AUTH and
+                    plugin_name is not None):
                 await self._process_auth(plugin_name, auth_packet)
             else:
                 # send legacy handshake
@@ -1076,7 +1794,7 @@ class Connection:
         pkt = await self._read_packet()
         pkt.check_error()
 
-    async def sha256_password_auth(self, pkt: Any) -> Any:
+    async def sha256_password_auth(self, pkt):
         if self._secure:
             logger.debug("sha256: Sending plain password")
             data = self._password.encode('latin1') + b'\0'
@@ -1118,19 +1836,19 @@ class Connection:
         return pkt
 
     # _mysql support
-    def thread_id(self) -> int:
+    def thread_id(self):
         return self.server_thread_id[0]
 
-    def character_set_name(self) -> str:
+    def character_set_name(self):
         return self._charset
 
-    def get_host_info(self) -> str:
+    def get_host_info(self):
         return self.host_info
 
-    def get_proto_info(self) -> Tuple[int, int, int]:
+    def get_proto_info(self):
         return self.protocol_version
 
-    async def _get_server_information(self) -> None:
+    async def _get_server_information(self):
         i = 0
         packet = await self._read_packet()
         data = packet.get_all_data()
@@ -1150,7 +1868,7 @@ class Connection:
 
         self.server_capabilities = struct.unpack('<H', data[i:i + 2])[0]
         i += 2
-
+        salt_len = None
         if len(data) >= i + 6:
             lang, stat, cap_h, salt_len = struct.unpack('<BHHB', data[i:i + 6])
             i += 6
@@ -1178,6 +1896,7 @@ class Connection:
         i += 1
 
         # AUTH PLUGIN NAME may appear here.
+        # noinspection PyUnresolvedReferences
         if self.server_capabilities & CLIENT.PLUGIN_AUTH and len(data) >= i:
             # Due to Bug#59453 the auth-plugin-name is missing the terminating
             # NUL-char in versions prior to 5.5.10 and 5.6.2.
@@ -1192,7 +1911,7 @@ class Connection:
             else:
                 self._server_auth_plugin = data[i:server_end].decode('latin1')
 
-    def get_transaction_status(self) -> bool:
+    def get_transaction_status(self):
         return bool(self.server_status & SERVER_STATUS.SERVER_STATUS_IN_TRANS)
 
     def get_server_info(self):
@@ -1200,18 +1919,18 @@ class Connection:
 
     # Just to always have consistent errors 2 helpers
 
-    def _close_on_cancel(self) -> None:
+    def _close_on_cancel(self):
         self.close()
         self._close_reason = "Cancelled during execution"
 
-    def _ensure_alive(self) -> None:
+    def _ensure_alive(self):
         if not self._writer:
             if self._close_reason is None:
                 raise InterfaceError("(0, 'Not connected')")
             else:
                 raise InterfaceError(self._close_reason)
 
-    def __del__(self) -> None:
+    def __del__(self):
         if self._writer:
             warnings.warn("Unclosed connection {!r}".format(self),
                           ResourceWarning)
@@ -1229,29 +1948,26 @@ class Connection:
     NotSupportedError = NotSupportedError
 
 
-async def _close_cursor(c: Cursor) -> None:
-    await c.close()
-
-
 # TODO: move OK and EOF packet parsing/logic into a proper subclass
 # of MysqlPacket like has been done with FieldDescriptorPacket.
 class MySQLResult:
 
-    def __init__(self, connection: Connection) -> None:
-        self.connection: Connection = connection
-        self.affected_rows: Optional[int] = None
-        self.insert_id: Optional[int] = None
-        self.server_status: Optional[int] = None
-        self.warning_count: int = 0
-        self.message: Optional[str] = None
-        self.field_count: int = 0
-        self.description: Optional[List] = None
-        self.rows: Optional[List] = None
-        self.has_next: Optional[bool] = None
-        self.unbuffered_active: bool = False
+    def __init__(self, connection):
+        self.connection = connection
+        self.affected_rows = None
+        self.insert_id = None
+        self.server_status = None
+        self.warning_count = 0
+        self.message = None
+        self.field_count = 0
+        self.description = None
+        self.rows = None
+        self.has_next = None
+        self.unbuffered_active = False
 
-    async def read(self) -> None:
+    async def read(self):
         try:
+            # noinspection PyProtectedMember
             first_packet = await self.connection._read_packet()
 
             # TODO: use classes for different packet types?
@@ -1264,8 +1980,9 @@ class MySQLResult:
         finally:
             self.connection = None
 
-    async def init_unbuffered_query(self) -> None:
+    async def init_unbuffered_query(self):
         self.unbuffered_active = True
+        # noinspection PyProtectedMember
         first_packet = await self.connection._read_packet()
 
         if first_packet.is_ok_packet():
@@ -1285,7 +2002,7 @@ class MySQLResult:
             # we set it to this instead of None, which would be preferred.
             self.affected_rows = 18446744073709551615
 
-    def _read_ok_packet(self, first_packet) -> None:
+    def _read_ok_packet(self, first_packet):
         ok_packet = OKPacketWrapper(first_packet)
         self.affected_rows = ok_packet.affected_rows
         self.insert_id = ok_packet.insert_id
@@ -1294,22 +2011,23 @@ class MySQLResult:
         self.message = ok_packet.message
         self.has_next = ok_packet.has_next
 
-    async def _read_load_local_packet(self, first_packet) -> None:
+    async def _read_load_local_packet(self, first_packet):
         load_packet = LoadLocalPacketWrapper(first_packet)
         sender = LoadLocalFile(load_packet.filename, self.connection)
         try:
             await sender.send_data()
         except Exception:
             # Skip ok packet
+            # noinspection PyProtectedMember
             await self.connection._read_packet()
             raise
-
+        # noinspection PyProtectedMember
         ok_packet = await self.connection._read_packet()
         if not ok_packet.is_ok_packet():
             raise OperationalError(2014, "Commands Out of Sync")
         self._read_ok_packet(ok_packet)
 
-    def _check_packet_is_eof(self, packet) -> bool:
+    def _check_packet_is_eof(self, packet):
         if packet.is_eof_packet():
             eof_packet = EOFPacketWrapper(packet)
             self.warning_count = eof_packet.warning_count
@@ -1317,16 +2035,16 @@ class MySQLResult:
             return True
         return False
 
-    async def _read_result_packet(self, first_packet) -> None:
+    async def _read_result_packet(self, first_packet):
         self.field_count = first_packet.read_length_encoded_integer()
         await self._get_descriptions()
         await self._read_rowdata_packet()
 
-    async def _read_rowdata_packet_unbuffered(self) -> Optional[Tuple[Optional[Any], ...]]:
+    async def _read_rowdata_packet_unbuffered(self):
         # Check if in an active query
         if not self.unbuffered_active:
             return
-
+        # noinspection PyProtectedMember
         packet = await self.connection._read_packet()
         if self._check_packet_is_eof(packet):
             self.unbuffered_active = False
@@ -1340,12 +2058,13 @@ class MySQLResult:
         self.rows = (row,)
         return row
 
-    async def _finish_unbuffered_query(self) -> None:
+    async def _finish_unbuffered_query(self):
         # After much reading on the MySQL protocol, it appears that there is,
         # in fact, no way to stop MySQL from sending all the data after
         # executing a query, so we just spin, and wait for an EOF packet.
         while self.unbuffered_active:
             try:
+                # noinspection PyProtectedMember
                 packet = await self.connection._read_packet()
             except OperationalError as e:
                 # TODO: replace these numbers with constants when available
@@ -1366,10 +2085,11 @@ class MySQLResult:
                 # release reference to kill cyclic reference.
                 self.connection = None
 
-    async def _read_rowdata_packet(self) -> None:
+    async def _read_rowdata_packet(self):
         """Read a rowdata packet for each data row in the result set."""
         rows = []
         while True:
+            # noinspection PyProtectedMember
             packet = await self.connection._read_packet()
             if self._check_packet_is_eof(packet):
                 # release reference to kill cyclic reference.
@@ -1380,7 +2100,7 @@ class MySQLResult:
         self.affected_rows = len(rows)
         self.rows = tuple(rows)
 
-    def _read_row_from_packet(self, packet) -> Tuple[Optional[Any], ...]:
+    def _read_row_from_packet(self, packet):
         row = []
         for encoding, converter in self.converters:
             try:
@@ -1397,7 +2117,7 @@ class MySQLResult:
             row.append(data)
         return tuple(row)
 
-    async def _get_descriptions(self) -> None:
+    async def _get_descriptions(self):
         """Read a column descriptor packet for each column in the result."""
         self.fields = []
         self.converters = []
@@ -1405,12 +2125,14 @@ class MySQLResult:
         conn_encoding = self.connection.encoding
         description = []
         for i in range(self.field_count):
+            # noinspection PyProtectedMember
             field = await self.connection._read_packet(
                 FieldDescriptorPacket)
             self.fields.append(field)
             description.append(field.description())
             field_type = field.type_code
             if use_unicode:
+                # noinspection PyUnresolvedReferences
                 if field_type == FIELD_TYPE.JSON:
                     # When SELECT from JSON column: charset = binary
                     # When SELECT CAST(... AS JSON): charset = connection
@@ -1436,7 +2158,7 @@ class MySQLResult:
             if converter is through:
                 converter = None
             self.converters.append((encoding, converter))
-
+        # noinspection PyProtectedMember
         eof_packet = await self.connection._read_packet()
         assert eof_packet.is_eof_packet(), 'Protocol error, expecting EOF'
         self.description = tuple(description)
@@ -1450,9 +2172,9 @@ class LoadLocalFile(object):
         self._file_object = None
         self._executor = None  # means use default executor
 
-    def _open_file(self) -> asyncio.Future:
+    def _open_file(self):
 
-        def opener(filename) -> None:
+        def opener(filename):
             try:
                 self._file_object = open(filename, 'rb')
             except IOError as e:
@@ -1462,7 +2184,7 @@ class LoadLocalFile(object):
         fut = self._loop.run_in_executor(self._executor, opener, self.filename)
         return fut
 
-    def _file_read(self, chunk_size: int) -> asyncio.Future:
+    def _file_read(self, chunk_size):
 
         def freader(chunk_size):
             try:
@@ -1482,8 +2204,9 @@ class LoadLocalFile(object):
         fut = self._loop.run_in_executor(self._executor, freader, chunk_size)
         return fut
 
-    async def send_data(self) -> None:
+    async def send_data(self):
         """Send data packets from the local file to the server"""
+        # noinspection PyProtectedMember
         self.connection._ensure_alive()
         conn = self.connection
 
@@ -1498,8 +2221,9 @@ class LoadLocalFile(object):
                     # TODO: consider drain data
                     conn.write_packet(chunk)
         except asyncio.CancelledError:
+            # noinspection PyProtectedMember
             self.connection._close_on_cancel()
-            raise
+            raise asyncio.CancelledError("send_data method was cancelled") from None
         finally:
             # send the empty packet to signify we are done sending data
             conn.write_packet(b"")
